+++ app/src/main/java/com/yenaly/han1meviewer/util/Networks.kt
@@ -1,14 +1,20 @@
 package com.yenaly.han1meviewer.util
 
 import com.google.common.util.concurrent.ListenableFuture
+import com.yenaly.han1meviewer.R
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.suspendCancellableCoroutine
 import okhttp3.Call
 import okhttp3.Callback
 import okhttp3.Response
 import java.io.IOException
+import java.net.ConnectException
+import java.net.SocketException
+import java.net.SocketTimeoutException
+import java.net.UnknownHostException
 import java.util.concurrent.ExecutionException
 import java.util.concurrent.Executor
+import javax.net.ssl.SSLHandshakeException
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 
@@ -83,5 +89,63 @@ private data object DirectExecutor : Executor {
 
     override fun execute(command: Runnable) {
         command.run()
+    }
+}
+
+/**
+ * 将首页加载异常映射为对应的错误提示字符串资源。
+ *
+ * 优先根据异常类型判断常见网络问题，必要时回退到异常信息中的关键字匹配。
+ *
+ * @receiver 首页加载过程中抛出的异常
+ * @return 错误提示的字符串资源 ID
+ */
+fun Throwable.toNetworkErrorMessageRes(): Int {
+    val rawMessage = message.orEmpty().lowercase()
+    return when {
+        this is UnknownHostException ||
+                rawMessage.contains("unable to resolve host") ||
+                rawMessage.contains("no address associated with hostname") -> {
+            R.string.home_error_dns
+        }
+
+        this is SocketTimeoutException || rawMessage.contains("timeout") -> {
+            R.string.home_error_timeout
+        }
+
+        this is SSLHandshakeException ||
+                rawMessage.contains("ssl") ||
+                rawMessage.contains("certificate") -> {
+            R.string.home_error_ssl
+        }
+
+        this is ConnectException || rawMessage.contains("failed to connect") -> {
+            R.string.home_error_connect
+        }
+
+        this is SocketException && rawMessage.contains("connection reset") -> {
+            R.string.home_error_connection_interrupted
+        }
+
+        rawMessage.contains("connection reset") -> {
+            R.string.home_error_connection_reset
+        }
+
+        rawMessage.contains("403") -> {
+            R.string.home_error_forbidden
+        }
+
+        rawMessage.contains("404") -> {
+            R.string.home_error_not_found
+        }
+
+        rawMessage.contains("500") || rawMessage.contains("502") ||
+                rawMessage.contains("503") || rawMessage.contains("504") -> {
+            R.string.home_error_server_unavailable
+        }
+
+        else -> {
+            R.string.home_error_generic
+        }
     }
 }
